Added gps nmea ubx i2c and modif of uart and enable of printf with syscalls

+++ inc/gpio.h
@@ -2,4 +2,4 @@
 
 #include <stdint.h>
 
-uint8_t init_gpios(void);
+int init_gpios(void);
